Rearranged header files (introducing nv2a_regs.h)
