+++ src/devices/video/nv2a_debug.h
@@ -25,9 +25,6 @@
 
 // Enable for NV2A Debug logging (Warning: Slow!)
 // #define DEBUG_NV2A
-#include "CxbxKrnl/gloffscreen/gloffscreen.h"
-#include "CxbxKrnl/gloffscreen/glextensions.h"
-
 #ifdef DEBUG_NV2A
 # define NV2A_DPRINTF(format, ...)       printf("[0x????] NV2A: " format, ## __VA_ARGS__)
 #else
@@ -39,6 +36,9 @@
 #ifdef DEBUG_NV2A_GL
 
 #include <stdbool.h>
+#include "CxbxKrnl/gloffscreen/gloffscreen.h"
+#include "CxbxKrnl/gloffscreen/glextensions.h"
+
 void gl_debug_message(bool cc, const char *fmt, ...);
 void gl_debug_group_begin(const char *fmt, ...);
 void gl_debug_group_end(void);
